+++ test/src/k-provider/ovp/provider.spec.js
@@ -39,6 +39,27 @@ describe('OVPProvider.partnerId:1082342', function() {
     );
   });
 
+  it('should apply the request host regex on the source urls', done => {
+    sinon.stub(MultiRequestBuilder.prototype, 'execute').callsFake(function() {
+      return new Promise(resolve => {
+        resolve({response: new MultiRequestResult(BE_DATA.AnonymousMocEntryWithRequestHostRegexAction.response)});
+      });
+    });
+    provider.getMediaConfig({entryId: '1_rsrdfext'}).then(
+      mediaConfig => {
+        try {
+          mediaConfig.should.deep.equal(MEDIA_CONFIG_DATA.RegexAppliedSources);
+          done();
+        } catch (err) {
+          done(err);
+        }
+      },
+      err => {
+        done(err);
+      }
+    );
+  });
+
   it('should return config with plugins and without drm data', done => {
     provider = new OVPProvider({partnerId: partnerId, uiConfId: 38621471}, playerVersion);
     sinon.stub(MultiRequestBuilder.prototype, 'execute').callsFake(function() {
@@ -99,6 +120,47 @@ describe('OVPProvider.partnerId:1082342', function() {
       },
       err => {
         done(err);
+      }
+    );
+  });
+
+  it('should return block error for server block response', done => {
+    sinon.stub(MultiRequestBuilder.prototype, 'execute').callsFake(function() {
+      return new Promise(resolve => {
+        resolve({response: new MultiRequestResult(BE_DATA.BlockActionEntry.response)});
+      });
+    });
+    provider.getMediaConfig({entryId: 1234}).then(
+      mediaConfig => {
+        try {
+          throw new Error('no error returned where block action error was expected', mediaConfig);
+        } catch (e) {
+          done(e);
+        }
+      },
+      err => {
+        const expected = {
+          severity: 2,
+          category: 2,
+          code: 2001,
+          data: {
+            action: {
+              type: 1
+            },
+            messages: [
+              {
+                message: "Un authorized country\nWe're sorry, this content is only available in certain countries.",
+                code: 'COUNTRY_RESTRICTED'
+              }
+            ]
+          }
+        };
+        try {
+          err.should.deep.equal(expected);
+          done();
+        } catch (e) {
+          done(e);
+        }
       }
     );
   });